RAG-knowledge/flows/chat_basic/requirements.txt
azure-search-documents==11.6.0b4
python-dotenv
openai
tiktoken